+++ client/src/fs/FsProvider.ts
@@ -1,21 +1,31 @@
-import * as vscode from "vscode"
 import { fromUri } from "../adt/AdtServer"
-import { FileSystemError, FileChangeType } from "vscode"
+import {
+  FileSystemError,
+  FileChangeType,
+  FileSystemProvider,
+  EventEmitter,
+  FileChangeEvent,
+  Event,
+  Uri,
+  Disposable,
+  FileStat,
+  FileType
+} from "vscode"
 
-export class FsProvider implements vscode.FileSystemProvider {
-  private pEventEmitter = new vscode.EventEmitter<vscode.FileChangeEvent[]>()
+export class FsProvider implements FileSystemProvider {
+  private pEventEmitter = new EventEmitter<FileChangeEvent[]>()
   // tslint:disable-next-line:member-ordering
-  public readonly onDidChangeFile: vscode.Event<vscode.FileChangeEvent[]> = this
-    .pEventEmitter.event
+  public readonly onDidChangeFile: Event<FileChangeEvent[]> = this.pEventEmitter
+    .event
 
   public watch(
-    uri: vscode.Uri,
+    uri: Uri,
     options: { recursive: boolean; excludes: string[] }
-  ): vscode.Disposable {
-    return new vscode.Disposable(() => undefined)
+  ): Disposable {
+    return new Disposable(() => undefined)
   }
 
-  public async stat(uri: vscode.Uri): Promise<vscode.FileStat> {
+  public async stat(uri: Uri): Promise<FileStat> {
     if (uri.path === "/.vscode") throw FileSystemError.FileNotFound(uri)
     const server = fromUri(uri)
     if (uri.path === "/") return server.findNode(uri)
@@ -26,23 +36,21 @@ export class FsProvider implements vscode.FileSystemProvider {
     }
   }
 
-  public async readDirectory(
-    uri: vscode.Uri
-  ): Promise<Array<[string, vscode.FileType]>> {
+  public async readDirectory(uri: Uri): Promise<Array<[string, FileType]>> {
     const server = fromUri(uri)
     const dir = server.findNode(uri)
     await server.refreshDirIfNeeded(dir)
     const contents = [...dir].map(
-      ([name, node]) => [name, node.type] as [string, vscode.FileType]
+      ([name, node]) => [name, node.type] as [string, FileType]
     )
     return contents
   }
-  public createDirectory(uri: vscode.Uri): void | Thenable<void> {
+  public createDirectory(uri: Uri): void | Thenable<void> {
     throw FileSystemError.NoPermissions(
       "Not a real filesystem, directory creation is not supported"
     )
   }
-  public async readFile(uri: vscode.Uri): Promise<Uint8Array> {
+  public async readFile(uri: Uri): Promise<Uint8Array> {
     const server = fromUri(uri)
     const file = server.findNode(uri)
 
@@ -54,7 +62,7 @@ export class FsProvider implements vscode.FileSystemProvider {
     throw FileSystemError.Unavailable(uri)
   }
   public async writeFile(
-    uri: vscode.Uri,
+    uri: Uri,
     content: Uint8Array,
     options: { create: boolean; overwrite: boolean }
   ): Promise<void> {
@@ -69,14 +77,14 @@ export class FsProvider implements vscode.FileSystemProvider {
     this.pEventEmitter.fire([{ type: FileChangeType.Changed, uri }])
   }
 
-  public async delete(uri: vscode.Uri, options: { recursive: boolean }) {
+  public async delete(uri: Uri, options: { recursive: boolean }) {
     const server = fromUri(uri)
     await server.delete(uri)
   }
 
   public rename(
-    oldUri: vscode.Uri,
-    newUri: vscode.Uri,
+    oldUri: Uri,
+    newUri: Uri,
     options: { overwrite: boolean }
   ): void | Thenable<void> {
     throw new Error("Method not implemented.")